requirements.txt
Django==4.2.7
numpy
opencv-python
matplotlib
Pillow
rembg
scikit-image
fpdf